refacto: several update

+++ src/main/java/com/openclassrooms/chatop/dto/rental/BaseRentalDto.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 abstract class BaseRentalDto {
@@ -15,11 +17,11 @@ abstract class BaseRentalDto {
 
     @NotNull
     @Positive
-    private Integer surface;
+    private Double surface;
 
     @NotNull
     @Positive
-    private Integer price;
+    private BigDecimal price;
 
     @NotNull
     private String description;
