+++ src/main/java/com/openclassrooms/chatop/dto/rental/RentalUpdateDto.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class RentalUpdateDto {
@@ -16,11 +18,11 @@ public class RentalUpdateDto {
 
     @NotNull
     @Positive
-    private Integer surface;
+    private Double surface;
 
     @NotNull
     @Positive
-    private Integer price;
+    private BigDecimal price;
 
     @NotBlank
     @NotNull
